+++ src/main/java/com/rampo/entity/ShopUserMap.java
@@ -36,7 +36,7 @@ public class ShopUserMap {
 	private boolean didRate;
 
 	@Column(name = "rating", nullable = true)
-	private float rating;
+	private long rating;
 
 	@Column(name = "didview", nullable = false)
 	private boolean didView;
